requirements.txt
requests
openai
python-dotenv
scikit-learn
tqdm
json-repair
